Service manager solution
Checkin of a service manager implementation
The AbstractServiceManager provides scope-based caching and services like IProvider instances

The ConfigFileSvcMgr provides service configuration via a JSON file. This file has a schema described in cfg/services.schema. A test config is in cfg/gui.services.json and stubbed-out interfaces are in the 4GlFanatics.UI* package and some basic implementations in Test.*

The 4GlFanatics.ServiceManager.IProvider interface is the way in which objects are instantiated. Implementations of this interface are in the 4GlFanatics.ServiceManager.Provider namespace.